Implemented different tile texture scales

+++ src/game/level.h
@@ -13,7 +13,7 @@ struct LevelData {
 	Uint32 img_tilesize;
 	Uint32 img_tilewidth;
 	Uint32 img_tilecount;
-	std::unique_ptr<Uint16[]> data;
+	std::unique_ptr<Uint32[]> data;
 
 	void load_from_file(const std::string& path);
